requirements.txt
streamlit
streamlit_drawable_canvas
pillow
psycopg2
python-dotenv